Move Authors and AuthorSpace under Settings namespace
