--- app/views/authors/show.html.erb
@@ -1,7 +1,0 @@
-<h1>Author Profile</h1>
-
-<p>Name: <%= @author.name %></p>
-<p>GitHub username: <%= @author.github_username %></p>
-
-<%= link_to 'Edit name', edit_author_path %>
-<%= link_to 'View repositories', author_repositories_path %>
