+++ app/views/settings/authors/edit.html.erb
@@ -3,4 +3,4 @@
 <%= render 'form', author: @author %>
 
 <br />
-<%= link_to 'Back to profile', author_path %>
+<%= link_to 'Back to profile', settings_author_path %>
