+++ app/views/settings/authors/repositories/_new_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: repository, url: '/author/repositories') do |form| %>
+<%= form_with(model: [:settings, :author, repository]) do |form| %>
   <%= render 'layouts/form_errors', resource: repository if repository.errors.any? %>
 
   <div>
