+++ app/views/settings/authors/repositories/new.html.erb
@@ -3,4 +3,4 @@
 <%= render 'new_form', repository: @repository %>
 
 <br />
-<%= link_to 'Back to repositories', author_repositories_path %>
+<%= link_to 'Back to repositories', settings_author_repositories_path %>
