+++ app/views/settings/authors/show.html.erb
@@ -1,0 +1,7 @@
+<h1>Author Profile</h1>
+
+<p>Name: <%= @author.name %></p>
+<p>GitHub username: <%= @author.github_username %></p>
+
+<%= link_to 'Edit name', edit_settings_author_path %>
+<%= link_to 'View repositories', settings_author_repositories_path %>
