+++ app/views/static_pages/index.html.erb
@@ -1,4 +1,3 @@
 <h1>HOME</h1>
 
-<%= link_to 'Author Profile', author_path %>
 <%= link_to 'Dashboard', dashboard_root_path %>
