+++ spec/systems/settings/authors/repositories/index_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe 'AuthorSpace::Repositories #index', type: :system do
     scenario "displays author's repositories" do
       page.set_rack_session(author_id: author.id)
 
-      visit author_repositories_path
+      visit settings_author_repositories_path
       expect(page).to have_content("#{author.name}'s Repositories")
       expect(page).to have_content(repo.name)
     end
@@ -17,14 +17,14 @@ RSpec.describe 'AuthorSpace::Repositories #index', type: :system do
     scenario 'does not display repositories by other authors' do
       page.set_rack_session(author_id: author.id)
 
-      visit author_repositories_path
+      visit settings_author_repositories_path
       expect(page).not_to have_content(repo_from_other_author.name)
     end
   end
 
   context 'when not signed in as an author' do
     scenario 'redirects to root path' do
-      visit author_repositories_path
+      visit settings_author_repositories_path
 
       expect(page).to have_current_path(root_path)
       expect(page).to have_content('Please sign in with GitHub.')
